LIBWHCA-140 Refactor logic
