+++ src/processing/management/commands/finalize_redactions.py
@@ -1,25 +1,11 @@
 import logging
 
-from bs4 import BeautifulSoup
-
 from django.core.management import BaseCommand
 
-from processing.models import Item
+from processing.common.finalize_redactions import convert_redaction
 
 
 logger = logging.getLogger(__name__)
-
-
-def redact_final(html):
-    soup = BeautifulSoup(html, 'lxml')
-    tags = soup.find_all('del', {'class': 'redacted'})
-
-    for tag in tags:
-        new_string = '▊▊▊▊▊▊▊▊▊▊'
-        tag = tag.replace_with(new_string)
-
-    html = str(soup)
-    return html
 
 
 class Command(BaseCommand):
@@ -29,14 +15,5 @@ class Command(BaseCommand):
         parser.add_argument('batch_selected', type=int, help='Batch to have redactions finalized.')
 
     def handle(self, *args, **options):
-        items = Item.objects.filter(batch=options['batch_selected'])
-        item = Item.objects.all()
-        for item in items:
-            logger.info(f'Marking: {item.id}, {item.title}')
-
-            html = item.body_redact
-
-            html = redact_final(html)
-
-            item.body_final = html
-            item.save(update_fields=['body_final'])
+        batch_selected = batch=options['batch_selected']
+        convert_redaction(batch_selected)
